pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.1",
    "appwrite>=13.4.1",
    "flask>=3.1.2",
    "pygithub>=2.8.1",
    "python-gitlab>=6.5.0",
    "resend>=2.17.0",
    "schedule>=1.2.2",
    "stripe>=13.0.1",
    "supabase>=2.22.2",
    "pytest>=8.4.2",
    "requests>=2.32.5",
    "openai>=1.58.1",
    "redis>=5.2.1",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.11",
    "email-validator>=2.3.0",
    "gunicorn>=23.0.0",
    "flask-login>=0.6.3",
    "sqlalchemy>=2.0.44",
    "werkzeug>=3.1.3",
    "pyjwt>=2.10.1",
]
